// packages/cactus-plugin-satp-hermes/src/main/typescript/core/errors/satp-error-type.ts
/**
 * SATP protocol error type classification.
 */
export enum SATPErrorType {
  UNSPECIFIED = 0,
  BADLY_FORMATED_MESSAGE = 1,
  INCORRECT_PARAMETER = 2,
  ACK_MISMATCH = 3,
  BADLY_FORMATED_MESSAGE_CLAIM = 4,
  BADLY_FORMATED_MESSAGE_BAD_SIGNATURE = 5,
  BADLY_FORMATED_MESSAGE_WRONG_TRANSACTION_ID = 6,
  BADLY_FORMATED_MESSAGE_MISMATCH_HASH_VALUES = 7,
  EXPIRED_SIGNING_KEY_CERTIFICATE = 8,
  EXPIRED_CLAIM = 9,
  MESSAGE_OUT_OF_SEQUENCE = 10,
  SESSION_NOT_FOUND = 11,
  SESSION_ID_NOT_FOUND = 12,
  PUBLIC_KEY_NOT_FOUND = 13,
  SENDER_GATEWAY_NETWORK_ID_NOT_FOUND = 14,
  COMMON_BODY_BADLY_FORMATED = 15,
  SESSION_MISS_MATCH = 16,
  SESSION_DATA_LOADED_INCORRECTLY = 17,
  SESSION_DATA_NOT_FOUND = 18,
  SESSION_COMPLETED = 19,
  MISSING_PARAMETER = 20,
  SATP_VERSION_NOT_SUPPORTED = 21,
  SIGNATURE_VERIFICATION_FAILED = 22,
  DLT_NOT_SUPPORTED = 23,
  HASH_MISS_MATCH = 24,
  CONTEXT_ID_MISS_MATCH = 25,
  BRIDGE_PROBLEM = 26,
  LOCK_ASSERTION_BADLY_FORMATED = 27,
  LOCK_ASSERTION_CLAIM_FORMAT_MISSING = 28,
  LOCK_ASSERTION_EXPIRATION_ERROR = 29,
  BURN_ASSERTION_BADLY_FORMATED = 30,
  MINT_ASSERTION_BADLY_FORMATED = 31,
  ASSIGNMENT_ASSERTION_BADLY_FORMATED = 32,
  ONTOLOGY_BADLY_FORMATED = 33,
  WRAP_ASSERTION_BADLY_FORMATED = 34,
}